[REF] Float into position component

+++ src/ECS/Systems/ControllableSystem.cpp
@@ -47,11 +47,11 @@ int controllable_system(World &world, sf::Event event)
             if (event.type == sf::Event::MouseButtonPressed && (MouseInput)event.mouseButton.button == i->shoot) {
                 if (!(drawable[index]->rect.x == 0 && drawable[index]->rect.y == 0 && drawable[index]->rect.x_size == 0
                         && drawable[index]->rect.y_size == 0)) {
-                    world.create_laser(GameObject::LASER, Vector2i{position[index]->pos.x + (drawable[index]->rect.x_size * drawable[index]->scale.x) + 1,
+                    world.create_laser(GameObject::LASER, Vector2f{position[index]->pos.x + (drawable[index]->rect.x_size * drawable[index]->scale.x) + 1.0,
                         position[index]->pos.y + ((drawable[index]->rect.y_size * drawable[index]->scale.y) / 2)}, Vector2i{15, 0},
                         0.04, world.getClock().getElapsedTime().asSeconds());
                 } else {
-                    world.create_laser(GameObject::LASER, Vector2i{position[index]->pos.x
+                    world.create_laser(GameObject::LASER, Vector2f{position[index]->pos.x
                             + (world.getResourcesManager().get_texture(drawable[index]->path).getSize().x * drawable[index]->scale.x) + 1,
                         position[index]->pos.y + ((drawable[index]->rect.y_size * drawable[index]->scale.y) / 2)}, Vector2i{15, 0},
                         0.04, world.getClock().getElapsedTime().asSeconds());
