+++ src/ECS/Systems/WaveSystem.cpp
@@ -15,12 +15,12 @@ static float elapsed_time = 0.0;
 
 int wave_system(World &world)
 {
-    int random_y = 0;
+    float random_y = 0;
 
     if (world.getClock().getElapsedTime().asSeconds() > (elapsed_time + 3.0)) {
         elapsed_time += 3.0;
         random_y = std::rand() % world.getWindow().getSize().y;
-        world.create_enemy(GameObject::ENEMY, Vector2i{900, random_y}, Vector2i{-2, 0}, 0.04, world.getClock().getElapsedTime().asSeconds());
+        world.create_enemy(GameObject::ENEMY, Vector2f{900.0, random_y}, Vector2i{-2, 0}, 0.04, world.getClock().getElapsedTime().asSeconds());
     }
     return (0);
 }
